+++ WalletWasabi/Extensions/NBitcoinExtensions.cs
@@ -4,6 +4,7 @@ using NBitcoin.Protocol;
 using NBitcoin.RPC;
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading;
@@ -389,6 +390,28 @@ namespace NBitcoin
 			}
 
 			return null;
+		}
+
+		private static NumberFormatInfo CurrencyNumberFormat = new NumberFormatInfo()
+		{
+			NumberGroupSeparator = " ",
+			NumberDecimalDigits = 0
+		};
+
+		private static string ToCurrency(this Money btc, string currency, decimal exchangeRate, bool lurkingWifeMode = false)
+		{
+			var dollars = exchangeRate * btc.ToDecimal(MoneyUnit.BTC);
+
+			return lurkingWifeMode
+				? $"### {currency}"
+				: exchangeRate == default
+					? $"??? {currency}"
+					: $"{dollars.ToString("N", CurrencyNumberFormat)} {currency}";
+		}
+
+		public static string ToUsdString(this Money btc, decimal usdExchangeRate, bool lurkingWifeMode = false)
+		{
+			return ToCurrency(btc, "USD", usdExchangeRate, lurkingWifeMode);
 		}
 	}
 }
